✨feat(querybar): SKFP-596 add an remoteComponent properties
