+++ src/store/remote/selector.ts
@@ -1,0 +1,6 @@
+import { RootState } from 'store/types';
+import { initialState } from 'store/remote/types';
+
+export type TRemoteProps = initialState;
+
+export const remoteSelector = (state: RootState) => state.remote;
