+++ src/store/remote/types.ts
@@ -1,0 +1,13 @@
+export enum RemoteComponentList {
+  MondoTree = 'mondoTree',
+  HPOTree = 'hpoTree',
+}
+
+export type initialState = {
+  [RemoteComponentList.MondoTree]: {
+    [value: string]: any;
+  };
+  [RemoteComponentList.HPOTree]: {
+    [value: string]: any;
+  };
+};
